Use syn to parse the attribute tokens
Attributes are allowed to be arbitrary unstructured Rust tokens so there is an
extra step to parse them into a structured representation -- what `macro_rules!`
calls a $meta or "meta item". Once structured, we don't need to explicitly
iterate tokens or rearrange bytes of the string to access the string literal.

Some other cleanups:

- Generally you want to format errors with "{}" when panicking in a proc macro
  because otherwise they get shown to the user in Debug form.

- Re-exporting validate_macros from compile_validate avoids the user having to
  write both of those crates in their Cargo.toml.

- Quasi-quoting simplifies the string to tokens to string to token stream dance.

- In my experience attribute names are typically lowercase.